Добавляет общую логику обновления
